Fixes known vulnerabilities per `npm audit`
TL;DR
-----

Brings this fork up to date on for packages with known vulns

Details
-------

I'm planning to use this plugin in an OutSystems mobile app and the
build was failing in MABS. The best hypothesis I had as to why it
as failing was that the critical vulnerabilities `npm` found were
preventing the build.

To validate that hypothesis, I forked the repository and run through
the repair process with `npm audit fix` and `npm audit fix --force`,
though who knows how I'll fair with the breaking changes.

This change includes all of the patches that `npm` things I need, so
I can see if that moves MABS along.